+++ src/main/java/ru/verlioka/cmf/appservices/suppie/dao/CustomerDao.java
@@ -1,0 +1,7 @@
+package ru.verlioka.cmf.appservices.suppie.dao;
+
+import ru.verlioka.cmf.appservices.suppie.models.CustomerEntity;
+import ru.verlioka.cmf.core.dao.generic.IGenericDao;
+
+public interface CustomerDao extends IGenericDao<CustomerEntity, Long> {
+}
