+++ src/main/java/ru/verlioka/cmf/appservices/suppie/dao/TariffDao.java
@@ -1,0 +1,7 @@
+package ru.verlioka.cmf.appservices.suppie.dao;
+
+import ru.verlioka.cmf.appservices.suppie.models.TariffEntity;
+import ru.verlioka.cmf.core.dao.generic.IGenericDao;
+
+public interface TariffDao extends IGenericDao<TariffEntity, Long> {
+}
