+++ src/main/java/ru/verlioka/cmf/appservices/suppie/services/CustomerService.java
@@ -1,8 +1,8 @@
-package ru.verlioka.cmf.appservices.yandexmapsample.services.suppie;
+package ru.verlioka.cmf.appservices.suppie.services;
 
-import ru.verlioka.cmf.appservices.yandexmapsample.controllers.suppie.json.BalanceReport;
-import ru.verlioka.cmf.appservices.yandexmapsample.controllers.suppie.json.PaymentsReport;
-import ru.verlioka.cmf.appservices.yandexmapsample.models.suppie.CustomerEntity;
+import ru.verlioka.cmf.appservices.suppie.controllers.json.BalanceReport;
+import ru.verlioka.cmf.appservices.suppie.controllers.json.PaymentsReport;
+import ru.verlioka.cmf.appservices.suppie.models.CustomerEntity;
 import ru.verlioka.cmf.core.services.generic.db.IGenericService;
 
 import java.util.List;
