--- src/main/java/ru/verlioka/cmf/appservices/yandexmapsample/dao/suppie/CustomerDao.java
@@ -1,7 +1,0 @@
-package ru.verlioka.cmf.appservices.yandexmapsample.dao.suppie;
-
-import ru.verlioka.cmf.appservices.yandexmapsample.models.suppie.CustomerEntity;
-import ru.verlioka.cmf.core.dao.generic.IGenericDao;
-
-public interface CustomerDao extends IGenericDao<CustomerEntity, Long> {
-}
