--- src/main/java/ru/verlioka/cmf/appservices/yandexmapsample/dao/suppie/TariffDao.java
@@ -1,7 +1,0 @@
-package ru.verlioka.cmf.appservices.yandexmapsample.dao.suppie;
-
-import ru.verlioka.cmf.appservices.yandexmapsample.models.suppie.TariffEntity;
-import ru.verlioka.cmf.core.dao.generic.IGenericDao;
-
-public interface TariffDao extends IGenericDao<TariffEntity, Long> {
-}
